+++ rush-plugins/rush-migrate-subspace-plugin/src/functions/addProjectToSubspace.ts
@@ -6,7 +6,7 @@ import { Colorize } from '@rushstack/terminal';
 import { IRushConfigurationProjectJson } from '@rushstack/rush-sdk/lib/api/RushConfigurationProject';
 import { enterNewProjectLocationPrompt, moveProjectPrompt } from '../prompts/project';
 import { RushConstants } from '@rushstack/rush-sdk';
-import { addProjectToRushConfiguration } from './updateRushConfiguration';
+import { addProjectToRushConfiguration, removeProjectFromRushConfiguration } from './updateRushConfiguration';
 import {
   getRushSubspacesConfigurationJsonPath,
   isExternalMonorepo,
@@ -73,13 +73,14 @@ export const addProjectToSubspace = async (
   targetMonorepoPath: string
 ): Promise<void> => {
   Console.debug(
-    `Adding project ${Colorize.bold(sourceProject.packageName)} to subspace ${Colorize.bold(
+    `Adding source project ${Colorize.bold(sourceProject.packageName)} to target subspace ${Colorize.bold(
       targetSubspace
     )}...`
   );
 
   let targetProjectFolderPath: string | undefined = `${targetMonorepoPath}/${sourceProject.projectFolder}`;
-  if (isExternalMonorepo(sourceMonorepoPath, targetMonorepoPath) || (await moveProjectPrompt())) {
+  const isExternal: boolean = isExternalMonorepo(sourceMonorepoPath, targetMonorepoPath);
+  if (isExternal || (await moveProjectPrompt())) {
     const sourceProjectFolderPath: string = `${sourceMonorepoPath}/${sourceProject.projectFolder}`;
     targetProjectFolderPath = await moveProjectToSubspaceFolder(
       sourceProjectFolderPath,
@@ -99,11 +100,6 @@ export const addProjectToSubspace = async (
     }
   }
 
-  /** WARN: Disabling different repositories for now.
-  addProjectToRushConfiguration(sourceProject, targetSubspace, targetProjectFolderPath);
-  removeProjectFromRushConfiguration(sourceProject, sourceMonorepoPath);
-   */
-
   const targetLegacySubspaceFolderPath: string = `${targetProjectFolderPath}/subspace`;
   if (FileSystem.exists(targetLegacySubspaceFolderPath)) {
     Console.debug(`Removing legacy subspace folder ${Colorize.bold(targetLegacySubspaceFolderPath)}...`);
@@ -111,13 +107,17 @@ export const addProjectToSubspace = async (
   }
 
   addProjectToRushConfiguration(sourceProject, targetSubspace, targetProjectFolderPath, targetMonorepoPath);
+  if (isExternal) {
+    removeProjectFromRushConfiguration(sourceProject, sourceMonorepoPath);
+  }
+
   if (sourceProject.subspaceName) {
     refreshSubspace(sourceProject.subspaceName, sourceMonorepoPath);
   }
 
   Console.success(
-    `Project ${Colorize.bold(
+    `Source project ${Colorize.bold(
       sourceProject.packageName
-    )} has been successfully added to subspace ${Colorize.bold(targetSubspace)}.`
+    )} has been successfully added to target subspace ${Colorize.bold(targetSubspace)}.`
   );
 };