+++ rush-plugins/rush-migrate-subspace-plugin/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - rush-migrate-subspace-plugin
 
-This log was last generated on Tue, 07 Jan 2025 10:49:32 GMT and should not be manually modified.
+This log was last generated on Thu, 13 Feb 2025 03:23:57 GMT and should not be manually modified.
+
+## 1.4.1
+Thu, 13 Feb 2025 03:23:57 GMT
+
+### Patches
+
+- Fixes rush.json removing project
+
+## 1.4.0
+Mon, 03 Feb 2025 05:39:03 GMT
+
+### Minor changes
+
+- Supports external repository subspace migration
 
 ## 1.3.0
 Tue, 07 Jan 2025 10:49:32 GMT